# docs/api/openapi/api/openapi.yaml
openapi: 3.1.0
info:
  title: Chronos API
  description: "Shift, leave, and sickness planning \u2014 FS Phase 1 implementation."
  version: 0.2.0
paths:
...
  /api/projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: list_projects_api_projects_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectOut'
                type: array
                title: Response List Projects Api Projects Get
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}:
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project_api_projects__project_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/shifts:
    get:
      tags:
      - shifts
      summary: List Shifts
      operationId: list_shifts_api_shifts_get
      parameters:
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      - name: period_start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period Start
      - name: period_end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Period End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftOut'
                title: Response List Shifts Api Shifts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - shifts
      summary: Create Shift
      operationId: create_shift_api_shifts_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShiftIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/reports/leave:
    get:
      tags:
      - reports
      summary: Leave Summary
      operationId: leave_summary_api_reports_leave_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveSummaryRow'
                title: Response Leave Summary Api Reports Leave Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/reports/leave.csv:
    get:
      tags:
      - reports
      summary: Leave Summary Csv
      operationId: leave_summary_csv_api_reports_leave_csv_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/reports/sickness:
    get:
      tags:
      - reports
      summary: Sickness Summary
      operationId: sickness_summary_api_reports_sickness_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveSummaryRow'
                title: Response Sickness Summary Api Reports Sickness Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/reports/shifts:
    get:
      tags:
      - reports
      summary: Shifts Summary
      operationId: shifts_summary_api_reports_shifts_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          format: date
          title: Start
      - name: end
        in: query
        required: true
        schema:
          type: string
          format: date
          title: End
      - name: team_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Team Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftSummaryRow'
                title: Response Shifts Summary Api Reports Shifts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssignmentOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        shift_id:
          type: string
          format: uuid
          title: Shift Id
        user_id:
          type: string
          format: uuid
          title: User Id
        status:
          type: string
          title: Status
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - id
      - shift_id
      - user_id
      - status
      - note
      title: AssignmentOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LeaveSummaryRow:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        user_name:
          type: string
          title: User Name
        vacation_days:
          type: integer
          title: Vacation Days
        sickness_days:
          type: integer
          title: Sickness Days
        other_days:
          type: integer
          title: Other Days
      type: object
      required:
      - user_id
      - user_name
      - vacation_days
      - sickness_days
      - other_days
      title: LeaveSummaryRow
...
    ProjectIn:
      properties:
        code:
          type: string
          maxLength: 64
          minLength: 1
          title: Code
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
      type: object
      required:
      - code
      - name
      title: ProjectIn
    ProjectOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - id
      - code
      - name
      - description
      title: ProjectOut
...
    ShiftIn:
      properties:
        team_id:
          type: string
          format: uuid
          title: Team Id
        service_date:
          type: string
          format: date
          title: Service Date
        shift_type:
          $ref: '#/components/schemas/ShiftType'
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        required_headcount:
          type: integer
          maximum: 20.0
          minimum: 1.0
          title: Required Headcount
          default: 1
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
      type: object
      required:
      - team_id
      - service_date
      - shift_type
      - start_time
      - end_time
      title: ShiftIn
    ShiftOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        team_id:
          type: string
          format: uuid
          title: Team Id
        service_date:
          type: string
          format: date
          title: Service Date
        shift_type:
          type: string
          title: Shift Type
        start_time:
          type: string
          format: time
          title: Start Time
        end_time:
          type: string
          format: time
          title: End Time
        required_headcount:
          type: integer
          title: Required Headcount
        project_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Project Id
        assignments:
          items:
            $ref: '#/components/schemas/AssignmentOut'
          type: array
          title: Assignments
          default: []
      type: object
      required:
      - id
      - team_id
      - service_date
      - shift_type
      - start_time
      - end_time
      - required_headcount
      - project_id
      title: ShiftOut
    ShiftSummaryRow:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        user_name:
          type: string
          title: User Name
        early_count:
          type: integer
          title: Early Count
        late_count:
          type: integer
          title: Late Count
        other_count:
          type: integer
          title: Other Count
      type: object
      required:
      - user_id
      - user_name
      - early_count
      - late_count
      - other_count
      title: ShiftSummaryRow
    ShiftType:
      type: string
      enum:
      - EARLY
      - LATE
      - OTHER
      title: ShiftType
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...